Adicionando testes de empresaService.

+++ src/main/java/com/example/fiscaliza/services/EmpresaService.java
@@ -2,6 +2,7 @@ package com.example.fiscaliza.services;
 
 import com.example.fiscaliza.models.Empresa;
 import com.example.fiscaliza.repository.EmpresaRepository;
+import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +15,7 @@ public class EmpresaService {
     @Autowired
     private EmpresaRepository repository;
 
-    public Integer save(Empresa empresa){
+    public Long save(Empresa empresa){
         repository.save(empresa);
         return empresa.getId();
     }
@@ -32,7 +33,13 @@ public class EmpresaService {
         return empresa;
     }
 
-    public void delete(Optional<Empresa> empresa){
-        repository.delete(empresa.get());
+    public void delete(Optional<Empresa> empresa) throws Exception{
+
+        if(findById(empresa.get().getId()).isPresent()) {
+            repository.delete(empresa.get());
+        } else {
+            throw new Exception("Empresa não encontrada com o ID informado!");
+        }
+
     }
 }
